Built project
Warning: A build error was located. Growl Notifications attempts to optimize the injector, but the minimizer takes care of that, which ends up with both of the clashing and creating a JS error on init. It was removed manually in this build, but a fix will need to be found for eventual ones